# Add a low-level array transfer performance test