--- a/test/performance/comm/low-level/array-gets.ml-perf.graph
+++ b/test/performance/comm/low-level/array-gets.ml-perf.graph
@@ -1,0 +1,11 @@
+perfkeys: MB/s: , MB/s:  
+files: array-get-512B.dat, array-get-4096B.dat
+graphkeys: 512B array, 4096B array
+graphtitle: Small Array GET Performance
+ylabel: Performance (MB/s)
+
+perfkeys: MB/s: , MB/s: , MB/s: 
+files: array-get-1M.dat, array-get-1G.dat, array-get-FR4.dat
+graphkeys: 1M array, 1G array, 1/4 mem array
+graphtitle: Large Array GET Performance
+ylabel: Performance (MB/s)