--- a/test/performance/comm/low-level/array-puts.ml-perf.graph
+++ b/test/performance/comm/low-level/array-puts.ml-perf.graph
@@ -1,0 +1,11 @@
+perfkeys: MB/s: , MB/s:  
+files: array-put-512B.dat, array-put-4096B.dat
+graphkeys: 512B array, 4096B array
+graphtitle: Small Array PUT Performance
+ylabel: Performance (MB/s)
+
+perfkeys: MB/s: , MB/s: , MB/s: 
+files: array-put-1M.dat, array-put-1G.dat, array-put-FR4.dat
+graphkeys: 1M array, 1G array, 1/4 mem array
+graphtitle: Large Array PUT Performance
+ylabel: Performance (MB/s)