--- a/test/performance/comm/low-level/arrayTransfer.ml-execopts
+++ b/test/performance/comm/low-level/arrayTransfer.ml-execopts
@@ -1,0 +1,11 @@
+--xferMem=512   --runSecs=1 --op=opPut # array-put-512B
+--xferMem=4096  --runSecs=1 --op=opPut # array-put-4096B
+--xferMB=1      --runSecs=1 --op=opPut # array-put-1M
+--xferMB=1024   --runSecs=1 --op=opPut # array-put-1G
+--memFraction=4 --runSecs=1 --op=opPut # array-put-FR4
+
+--xferMem=512   --runSecs=1 --op=opGet # array-get-512B
+--xferMem=4096  --runSecs=1 --op=opGet # array-get-4096B
+--xferMB=1      --runSecs=1 --op=opGet # array-get-1M
+--xferMB=1024   --runSecs=1 --op=opGet # array-get-1G
+--memFraction=4 --runSecs=1 --op=opGet # array-get-FR4