This adds perf testing for low-ish level communication: large gets/puts that
are initiated as part of an array bulk transfer. This does a point-to-point
transfer for small (512B, 4096B) and large (1MB, 1GB, 1/4 mem) array transfers.

This test will be useful to determine the threshold for BTE transfers and also
shows that ugni sustained bandwidth is lagging because it does not use BTE for
large transfers yet. For 1MB and larger transfers ugni only achieves ~3 GB/s,
while gn-aries achieves ~8 GB/s. The sustained bandwidth for aries is somewhere
between 8-10 GB/s for unidirectional point-to-point puts/gets, so 8 GB/s is
probably pretty close to optimal. (Our bulk transfer includes some operations
and comm beyond the large get/put, so we shouldn't expect to quite see the
hardware limit in this test.)

Below is a table of results for the put version comparing ugni to gn-aries at
different transfer sizes. ugni is slightly ahead for small transfers, but is
over 2x behind for larger transfers that test sustained bandwidth. ugni maxes
out at around 3.7 GB/s for puts (2.7 for gets) and gn-aries maxes out at around
8.5 GB/s for both puts/gets.

| size    | ugni-put   | gn-put    |
| ------- | ---------- | --------- |
|  512B   |   35 MB/s  |   28 MB/s |
| 4096B   |  290 MB/s  |  228 MB/s |
|    1MB  | 3650 MB/s  | 7956 MB/s |
|    1GB  | 3707 MB/s  | 8552 MB/s |
| 1/4 mem | 3799 MB/s  | 8696 MB/s |


Related to https://github.com/chapel-lang/chapel/issues/8305 and https://github.com/chapel-lang/chapel/issues/9615