@ronawho 2018-05-30
This test is an amalgamation of performance/comm/low-level/many-to-one and runtime/configMatters/comm/bigTransfer